Reconstruction stencils complete.
Reconstruction stencils are now generated using vector valued rational approximation,
and underlying symmetries within each (sub)stencil are utilized as much as possible.

Stencil values are re-normalized to be exact for constants after being formed. This
seems to be necessary, but should be examined in more detail later.

Finally, the boundary conditions were a red herring for the normalization issue
mentioned above. The function to write data files including the halo will stay
in the code should it be useful for other debugging tasks in the future.